Move Project entities inside own namespace
