+++ src/Repository/Project/ProjectRepository.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Repository;
+namespace App\Repository\Project;
 
 use App\Entity\Accounting\Accounting;
-use App\Entity\Project;
+use App\Entity\Project\Project;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
